+++ examples/SearchInput/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { storiesOf } from '@kadira/storybook';
+import { storiesOf } from '@storybook/react';
 
 // For props table
 import SearchInput, { PureSearchInput } from 'src/SearchInput';